# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User API",
        "description": "API documentation for User management",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/bus-schedules/{id}": {
            "get": {
                "tags": [
                    "BusSchedule"
                ],
                "summary": "Get a specific bus schedule by ID",
                "operationId": "1d6838a88b7e1a1fa4a00091b5c8d86d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the bus schedule to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully fetched the bus schedule",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "bus_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "start_from": {
                                            "type": "string",
                                            "example": "Cumilla"
                                        },
                                        "date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2024-12-15"
                                        },
                                        "time": {
                                            "type": "string",
                                            "format": "time",
                                            "example": "12:30"
                                        },
                                        "total_tickets": {
                                            "type": "integer",
                                            "example": 50
                                        },
                                        "ticket_sold": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "remaining_tickets": {
                                            "type": "integer",
                                            "example": 40
                                        },
                                        "price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 250.5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bus schedule not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Bus schedule not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Something went wrong"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "BusSchedule"
                ],
                "summary": "Update an existing bus schedule",
                "operationId": "38cab7faa96da6b86b36f79cd669384b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the bus schedule to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "bus_id",
                                    "start_from",
                                    "date",
                                    "time",
                                    "total_tickets",
                                    "ticket_sold",
                                    "remaining_tickets",
                                    "price"
                                ],
                                "properties": {
                                    "id": {
                                        "description": "ID of the bus schedule",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "bus_id": {
                                        "description": "ID of the bus",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "start_from": {
                                        "description": "Location where the bus starts",
                                        "type": "string",
                                        "example": "Cumilla"
                                    },
                                    "date": {
                                        "description": "Date of the bus schedule",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-12-15"
                                    },
                                    "time": {
                                        "description": "Time when the bus departs",
                                        "type": "string",
                                        "format": "time",
                                        "example": "12:30"
                                    },
                                    "total_tickets": {
                                        "description": "Total number of tickets available for the bus schedule",
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "ticket_sold": {
                                        "description": "Number of tickets already sold",
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "remaining_tickets": {
                                        "description": "Remaining tickets for the bus schedule",
                                        "type": "integer",
                                        "example": 40
                                    },
                                    "price": {
                                        "description": "Price per ticket for the bus schedule",
                                        "type": "number",
                                        "format": "float",
                                        "example": 250.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully updated the bus schedule",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "bus_id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "start_from": {
                                            "type": "string",
                                            "example": "Cumilla"
                                        },
                                        "date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2024-12-15"
                                        },
                                        "time": {
                                            "type": "string",
                                            "format": "time",
                                            "example": "12:30"
                                        },
                                        "total_tickets": {
                                            "type": "integer",
                                            "example": 50
                                        },
                                        "ticket_sold": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "remaining_tickets": {
                                            "type": "integer",
                                            "example": 40
                                        },
                                        "price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 250.5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bus schedule not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Bus schedule not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request, validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Validation errors"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Something went wrong"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "BusSchedule"
                ],
                "summary": "Delete a bus schedule by ID",
                "operationId": "eafd20150d23b2132858045e3a0058a2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the bus schedule to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted the bus schedule",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bus schedule deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bus schedule not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Bus schedule not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Something went wrong"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}